BOOKKEEPER-572 Write to the journal before writing to the ledger storage.
Added a bound on the journal queue to stop it being overloaded. By default this is 200000, which should be enough to cover our current usecases. There were two tests which needed modifications. SlowBookieTest needed to wait for the bookkeeper client to finish processing so that all fragments would be complete. LedgerCacheTest needed to wait for the addEntry calls to complete, as the the order had changed.