Made ledger metadata immutable
Ensembles are now ImmutableMap<Long, ImmutableList<BookieSocketAddress>>.
The LedgerManager manager interface has changed to avoid implicitly
updating the version of a LedgerMetadata object. Recovery, Closing and
handling bookie failures have now changed, as there are no longer merge
operations. If a write to zookeeper fails, then
recovery/closing/failureHandling are retried with the new metadata,
provided that assumptions are not broken.

The interesting changes are in LedgerHandle and LedgerMetadata.

There are a lot of changes from ArrayList -> ImmutableList,List
Also a lot of changes from currentEnsemble -> getCurrentEnsemble.

One test had to change. With the old code, if someone fenced a ledger,
closing would fail. This isn't always the case now. If there's lac of
the writing ledger matches the lastEntryId of the new metadata, then
close can succeed. ConditionalSetTest has changed to reflect this.

Change-Id: I01c8c9a532f0c8802120acf93b8aeba20e2a3748