PaxosClient
Given a ledger, allows you to propose values for registers on that
ledger. The call will not complete until values have been agreed
on. It will keep trying to update until complete. The returned values
may not be those originally proposed but those proposed by another
instance of the client.