Increase timeouts in TestDisableEnsembleChange
In some narrow cases, if we manage to connect to the bookie, but the bookie
hasn't fully initialized, an add entry message will be dropped, and the
client will wait for the timeout to complete. The timeout is 5 seconds
by default, which was 5x the timeout in the test. This patch increases
the test timeout to 10s, which should give enough time for the addEntry
to fail and be retried.

Author: Ivan Kelly <ivan@midokura.com>

Reviewers: Enrico Olivelli <eolivelli@gmail.com>, Sijie Guo <sijie@apache.org>

This closes #646 from ivankelly/ensemble-change-test