ISSUE #691: Move generated files into their own module (#49)
Upcoming changes are using GRpc in conjunction with protobuf. GRpc
generates code that created deprecation warnings when compiled with
java 8, so this change moves all generated code out to another module,
so that we don't have to turn off -Werror for all code.

In any case, at some point we should split the bookkeeper client out
from the server module, at which point we would need the definitions
on a common place.